sorts properly
The heapsort function had a algorithm bug, and a implementation bug,
both of which escaped notice due to a critical design flaw in the test
helper ensureIntSlicesMatch function.

Heap sort was returning a min heap rather than sorting the value
list. But even that had a subtle implementation bug in it.

The fix was to first fix the test helper.  Then discovered that the
algorithm should be multi-stage, creating a max-heap in stage one then
sorting that max-heap in ascending order during stage 2.  Once that
was done I found a bug in creating the max-heap.  Once that was
corrected, all tests passed, and I added a few more tests to run many
permutations over random values to try to find a bug.